requirements.txt
asgiref==3.8.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==2.0.12
Django==3.2
django-extensions==3.2.3
django-filter==22.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.2
flake8==7.1.1
flake8-builtins==2.5.0
flake8-commas==4.0.0
flake8-docstrings==1.6.0
flake8-isort==6.1.2
flake8-multiline-containers==0.0.19
flake8-polyfill==1.0.2
flake8-quotes==3.4.0
flake8-variables-names==0.0.6
idna==3.10
iniconfig==2.0.0
isort==5.13.2
mccabe==0.7.0
packaging==24.2
pluggy==0.13.1
py==1.11.0
pycodestyle==2.12.1
pydocstyle==6.3.0
pyflakes==3.2.0
PyJWT==2.1.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
pytz==2024.2
requests==2.26.0
snowballstemmer==2.2.0
sqlparse==0.5.2
toml==0.10.2
typing_extensions==4.12.2
urllib3==1.26.20
